backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-image==0.22.0
scipy==1.11.4
pandas==2.1.3
tensorflow==2.14.0
keras==2.14.0
matplotlib==3.8.2
torch==2.1.1
torchvision==0.16.1
pydantic==2.5.0
python-dotenv==1.0.0
cors==1.0.1
requests==2.31.0
